requirements.txt
Django~=3.0.7
pyjwt
passlib